+++ pyaud/__init__.py
@@ -1,6 +1,6 @@
 """Framework for writing Python package audits."""
 from . import config, exceptions, plugins
-from ._environ import Environ, environ
+from ._environ import environ
 from ._indexing import files
 from ._main import main
 from ._objects import BasePlugin
@@ -15,7 +15,6 @@ from ._version import __version__
 
 __all__ = [
     "BasePlugin",
-    "Environ",
     "__version__",
     "config",
     "environ",